refactor csr interface to not require an instance of crypto
The csr methods (create_csr, get_csr_builder) parameter lists require a self&
argument, which they do not access to. This makes the interface awkward to use:
when you want to create a CSR you have to create a Crypto object from a private
key and certificate, only to pass them in again.

This instead splits the operation: it provides the functionality with two free
functions and provide a convenience interface on top that creates CSRs for
existing Crypto objects.